migrate the epacamd_eia crosswalk code into the cems module
